+++ lib/fxdart.dart
@@ -24,7 +24,17 @@ export 'src/async_iterable.dart'
         concurrentAsync,
         concurrentPoolAsync;
 export 'src/config.dart' show FxDart, FxConfig;
-export 'src/fx.dart' show fx, fxAsync, fxStream, Fx, FxAsync, FxNum, FxAsyncNum;
+export 'src/fx.dart'
+    show
+        fx,
+        fxAsync,
+        fxStream,
+        Fx,
+        FxAsync,
+        FxNum,
+        FxAsyncNum,
+        FxPair,
+        FxAsyncPair;
 export 'src/dart_aliases.dart';
 export 'src/lazy/combine.dart';
 export 'src/lazy/effect.dart';